+++ src/app/page.tsx
@@ -6,11 +6,11 @@ export const metadata: Metadata = {
 };
 
 const Hub: React.FC = () => (
-  <div className="m-auto max-w-7xl">
+  <main>
     <h1 className="max-w-[50%] text-7xl text-white">
       Discover and Empower with Monitoring Solutions.
     </h1>
-  </div>
+  </main>
 );
 
 export default Hub;