+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Unreleased
 ### Added
 - Satoshi Font, main navigation, and test for navigation functionality [OEMC-18](https://vizzuality.atlassian.net/browse/OEMC-18)
 - Map set up and layers display [OEMC-23](https://vizzuality.atlassian.net/jira/software/c/projects/OEMC/boards/95?selectedIssue=OEMC-23)
-
+- Map controls to move around the map (zoom in, out, fit bounds)[OEMC-44](https://vizzuality.atlassian.net/browse/OEMC-44?atlOrigin=eyJpIjoiMGYzNzk5OWQwNmExNDRkNjllOWE5NWMxNjc4MmIwMmQiLCJwIjoiaiJ9)
 
 
 ### Changed